requirements.txt
opencv-python==4.8.1.78
deepface==0.0.79
supabase==2.0.2
paho-mqtt==1.6.1
python-dotenv==1.0.0
numpy==1.24.3
tensorflow==2.13.0
requests==2.31.0
